requirements.txt
aniso8601==7.0.0
arabic-reshaper==2.1.3
asgiref==3.4.1
certifi==2021.5.30
cffi==1.14.6
charset-normalizer==2.0.4
cryptography==3.4.8
defusedxml==0.7.1
diff-match-patch==20200713
dj-database-url==0.5.0
Django==3.2.6
django-allauth==0.45.0
django-cors-headers==3.7.0
django-dropbox-storage==0.4.1
django-filter==2.4.0
django-heroku==0.0.0
django-import-export==2.5.0
django-storages==1.11.1
django-widget-tweaks==1.4.8
djangorestframework==3.12.4
dropbox==11.18.0
et-xmlfile==1.1.0
future==0.18.2
graphene==2.1.9
graphene-django==2.15.0
graphql-core==2.3.2
graphql-relay==2.0.1
gunicorn==20.1.0
html5lib==1.1
idna==3.2
Markdown==3.3.4
MarkupPy==1.14
oauthlib==3.1.1
odfpy==1.4.1
openpyxl==3.0.7
Pillow==8.3.1
ply==3.11
promise==2.3
psycopg2-binary==2.9.1
pycparser==2.20
PyJWT==2.1.0
PyPDF2==1.26.0
python-bidi==0.4.2
python-memcached==1.59
python3-openid==3.2.0
pytz==2021.1
PyYAML==5.4.1
reportlab==3.6.1
requests==2.26.0
requests-oauthlib==1.3.0
Rx==1.6.1
singledispatch==3.6.2
six==1.16.0
sqlparse==0.4.1
stone==3.2.1
tablib==3.0.0
text-unidecode==1.3
urllib3==1.26.6
webencodings==0.5.1
whitenoise==5.3.0
xhtml2pdf==0.2.5
xlrd==2.0.1
xlwt==1.3.0
aniso8601==7.0.0
arabic-reshaper==2.1.3
asgiref==3.4.1
certifi==2021.5.30
cffi==1.14.6
charset-normalizer==2.0.4
cryptography==3.4.8
defusedxml==0.7.1
diff-match-patch==20200713
dj-database-url==0.5.0
Django==3.2.6
django-allauth==0.45.0
django-cors-headers==3.7.0
django-dropbox-storage==0.4.1
django-filter==2.4.0
django-heroku==0.0.0
django-import-export==2.5.0
django-storages==1.11.1
django-widget-tweaks==1.4.8
djangorestframework==3.12.4
dropbox==11.18.0
et-xmlfile==1.1.0
future==0.18.2
graphene==2.1.9
graphene-django==2.15.0
graphql-core==2.3.2
graphql-relay==2.0.1
gunicorn==20.1.0
html5lib==1.1
idna==3.2
Markdown==3.3.4
MarkupPy==1.14
oauthlib==3.1.1
odfpy==1.4.1
openpyxl==3.0.7
Pillow==8.3.1
ply==3.11
promise==2.3
psycopg2-binary==2.9.1
pycparser==2.20
PyJWT==2.1.0
PyPDF2==1.26.0
python-bidi==0.4.2
python-memcached==1.59
python3-openid==3.2.0
pytz==2021.1
PyYAML==5.4.1
reportlab==3.6.1
requests==2.26.0
requests-oauthlib==1.3.0
Rx==1.6.1
singledispatch==3.6.2
six==1.16.0
sqlparse==0.4.1
stone==3.2.1
tablib==3.0.0
text-unidecode==1.3
urllib3==1.26.6
webencodings==0.5.1
whitenoise==5.3.0
xhtml2pdf==0.2.5
xlrd==2.0.1
xlwt==1.3.0
